back/requirements.txt
python-multipart
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
PyJWT
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
asyncpg==0.30.0
openai==1.40.3
fastapi-mail==1.4.1
jinja2==3.1.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
aiosqlite==0.19.0
greenlet==3.1.1
